pyproject.toml
[tool.poetry]
name = "evaluate-llms"
version = "0.1.0"
description = ""
authors = ["Tomoya Kanda <tomoya.kanda.2022b@mlab.info>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.0.350"
python-dotenv = "^1.0.0"
datasets = "^2.15.0"
openai = "^1.3.9"
argparse = "^1.4.0"
matplotlib = "^3.8.2"
jupyter = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
